+++ addons/godot_xterm/native/src/terminal.cpp
@@ -304,13 +304,24 @@ int Terminal::_draw_cb(struct tsm_screen* con,
                 term->font_size,
                 fgcol);
 
+        float underline_thickness = 0.0f;
+        float underline_position = 0.0f;
+        if (attr->underline) {
+            underline_thickness = MAX(1.0f, term->fonts[font_type]->get_underline_thickness(term->font_size));
+            underline_position = term->fonts[font_type]->get_underline_position(term->font_size);
+            float underline_y = cell_position.y + term->font_offset + underline_position;
+            Vector2 underline_start = Vector2(cell_position.x, underline_y);
+            Vector2 underline_end = Vector2(cell_position.x + width * term->cell_size.x, underline_y);
+            term->rs->canvas_item_add_line(term->char_canvas_item, underline_start, underline_end, fgcol, underline_thickness);
+        }
+
         // Draw directly to the main canvas item when doing a full redraw, as the fore_canvas_item will be hidden.
-        // Attribute effects need to be applied manually (replicating shader logic).
         if (term->framebuffer_age == 0) {
             if (attr->blink) {
                 // Currently blink characters are always visible on a full redraw.
             }
 
+            // Attribute effects need to be applied manually (replicating shader logic).
             if (attr->inverse && term->inverse_mode == InverseMode::INVERSE_MODE_INVERT) {
                 fgcol = Color(1.0 - fgcol.r, 1.0 - fgcol.g, 1.0 - fgcol.b, fgcol.a);
             }
@@ -321,6 +332,13 @@ int Terminal::_draw_cb(struct tsm_screen* con,
                     static_cast<uint64_t>(*ch),
                     term->font_size,
                     fgcol);
+
+            if (attr->underline) {
+                float underline_y = cell_position.y + term->font_offset + underline_position + term->style_normal->get_margin(SIDE_TOP);
+                Vector2 underline_start = Vector2(cell_position.x + term->style_normal->get_margin(SIDE_LEFT), underline_y);
+                Vector2 underline_end = Vector2(cell_position.x + width * term->cell_size.x + term->style_normal->get_margin(SIDE_LEFT), underline_y);
+                term->rs->canvas_item_add_line(term->get_canvas_item(), underline_start, underline_end, fgcol, underline_thickness);
+            }
         }
     }