+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/lihop/godot-xterm/compare/v4.0.1...HEAD)
 
+### Added
+
+- Support for the <u>underline</u> attribute (`\u001b[4m`).
+
 ## [v4.0.1](https://github.com/lihop/godot-xterm/compare/v4.0.0...v4.0.1) - 2025-10-17
 
 ### Fixed